requirements.txt
langchain==0.1.0
openai==0.28.1
python-dotenv==1.0.0
supabase==2.0.0
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
twilio==8.10.0
python-multipart==0.0.6 